requirements.txt
# Grok Voice AI Companion - Python Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# WebSocket support
websockets>=12.0

# HTTP client for API calls
httpx>=0.26.0
aiohttp>=3.9.0

# Environment variables
python-dotenv>=1.0.0

# Static files and templates
jinja2>=3.1.3
aiofiles>=23.2.1

# Audio processing (optional, for advanced features)
# numpy>=1.26.0
# sounddevice>=0.4.6

# Development
watchfiles>=0.21.0
